[JUnit] Always fire TestStarted/Finished for pickle (#1765)
Fixes a number of issues w.r.t. JUnit test events:

 1. Only fires events for atomic tests (i.e. pickles).

 2. Ensures pickle started/finished events are fired even when duplicate
    step definitions are encountered.
